Seperating Author related methods into IAuthorRepository and AuthorRepository
