+++ src/Chirp.Infrastructure/CheepRepository.cs
@@ -1,28 +1,14 @@
 using System;
 using Chirp.Core;
+
 public class CheepRepository : ICheepRepository
 {
     readonly ChirpContext db;
+    readonly AuthorRepository authorRepo;
     public CheepRepository(ChirpContext chirpContext)
     {
         db = chirpContext;
-    }
-
-    public void AddAuthor(string name, string email)
-    {
-        if (!db.Authors.Any(a => a.Email == email))
-        {
-            db.Add(new Author { Name = name, Email = email, Cheeps = new List<Cheep>() });
-        }
-        db.SaveChanges();
-        db.ChangeTracker.Clear();
-    }
-
-    public void DeleteAuthor(Author author)
-    {
-        db.Remove(author);
-        db.SaveChanges();
-        db.ChangeTracker.Clear();
+        authorRepo = new AuthorRepository(chirpContext);
     }
 
     public void WriteCheep(string text, DateTime publishTimestamp, AuthorDTO author)
@@ -36,7 +22,7 @@ public class CheepRepository : ICheepRepository
         }
         else
         {
-            AddAuthor(author.Name, author.Email);
+            authorRepo.AddAuthor(author.Name, author.Email);
             var getNewAuthor = db.Authors.FirstOrDefault(a => a.Email == author.Email);
             db.Add<Cheep>(new Cheep { Text = text, TimeStamp = publishTimestamp, AuthorId = getNewAuthor.AuthorId, Author = getNewAuthor });
 
@@ -45,9 +31,9 @@ public class CheepRepository : ICheepRepository
         db.ChangeTracker.Clear();
     }
 
-    public void DeleteCheep(Cheep cheep)
+    public void DeleteCheep(CheepDTO cheep)
     {
-        var cheepToDelete = db.Cheeps.Find(cheep.CheepId);
+        var cheepToDelete = db.Cheeps.Find(cheep); // Might be an issue here? .Find(cheep) of type CheepDTO
         if (cheepToDelete != null)
         {
             db.Cheeps.Remove(cheepToDelete);
@@ -66,52 +52,9 @@ public class CheepRepository : ICheepRepository
         .ToList();
         foreach (Cheep cheep in cheeps)
         {
-            cheepDTOList.Add(new CheepDTO(GetAuthor(cheep.AuthorId).Name, cheep.Text, cheep.TimeStamp.ToString()));
+            cheepDTOList.Add(new CheepDTO(authorRepo.GetAuthorByID(cheep.AuthorId).Name, cheep.Text, cheep.TimeStamp.ToString()));
         }
         return cheepDTOList;
-    }
-
-    public AuthorDTO GetAuthor(int id)
-    {
-        var author = db.Authors
-        .Where(b => b.AuthorId == id)
-        .FirstOrDefault();
-        if (author != null)
-        {
-            return new AuthorDTO(author.Name, author.Email);
-        }
-        else
-        {
-            throw new NullReferenceException("Author not found");
-        }
-
-    }
-
-    public AuthorDTO GetAuthor(string EmailOrName)
-    {
-        Author? author = null;
-        if (EmailOrName.Contains("@") == true)
-        {
-            author = db.Authors
-           .Where(b => b.Email == EmailOrName)
-           .FirstOrDefault();
-        }
-        else
-        {
-            author = db.Authors
-           .Where(b => b.Name == EmailOrName)
-           .FirstOrDefault();
-        }
-
-        if (author != null)
-        {
-            return new AuthorDTO(author.Name, author.Email);
-        }
-        else
-        {
-            throw new NullReferenceException("Author not found");
-        }
-
     }
 
     public CheepDTO GetCheepById(int id)
@@ -137,7 +80,7 @@ public class CheepRepository : ICheepRepository
         .ToList();
         foreach (Cheep cheep in cheeps)
         {
-            cheepDTOList.Add(new CheepDTO(GetAuthor(cheep.AuthorId).Name, cheep.Text, cheep.TimeStamp.ToString()));
+            cheepDTOList.Add(new CheepDTO(authorRepo.GetAuthorByID(cheep.AuthorId).Name, cheep.Text, cheep.TimeStamp.ToString()));
 
         }
         return cheepDTOList;
