+++ src/Chirp.Web/Pages/Public.cshtml
@@ -19,6 +19,13 @@
             </form>
         </div>
     }
+    else
+    {
+        <div class="cheepbox">
+            <h3>Login to cheep!</h3>
+        </div>
+
+    }
 
 
     @if (Model.Cheeps.Any())
